requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
grequests==0.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
PyYAML==5.2
requests==2.22.0
SQLAlchemy==1.3.12
urllib3==1.25.7
waitress==1.3.1
Werkzeug==0.16.0
